Use 64 bit shares to allow for 32 bit MPC calculations
See SPDZ2k paper. We're using k=32 and s=32.

+++ src/NiMPC/Triples.nim
@@ -6,5 +6,4 @@ type Triple = tuple[a, b, c: Secret]
 proc triple*(party: Party): Triple =
   result.a = Secret(party: party, share: 1)
   result.b = Secret(party: party, share: 1)
-  result.c = Secret(party: party, share: uint32(party.peers.len + 1))
-  
+  result.c = Secret(party: party, share: uint64(party.peers.len + 1))
